Struct glutin_app::window::Window

pub struct Window {
    kind: WindowKind,
    mouse_down_button: Cell<Option<MouseButton>>,
    mouse_down_point: Cell<Point2D<i32>>,
    event_queue: RefCell<Vec<WindowEvent>>,
    mouse_pos: Cell<Point2D<i32>>,
    key_modifiers: Cell<KeyModifiers>,
    current_url: RefCell<Option<ServoUrl>>,
    pending_key_event_char: Cell<Option<char>>,
    pressed_key_map: RefCell<Vec<(ScanCode, char)>>,
}

The type of a window.

Fields
kind: WindowKind mouse_down_button: Cell<Option<MouseButton>> mouse_down_point: Cell<Point2D<i32>> event_queue: RefCell<Vec<WindowEvent>> mouse_pos: Cell<Point2D<i32>> key_modifiers: Cell<KeyModifiers> current_url: RefCell<Option<ServoUrl>> pending_key_event_char: Cell<Option<char>>

The contents of the last ReceivedCharacter event for use in a subsequent KeyEvent.

pressed_key_map: RefCell<Vec<(ScanCode, char)>>

The list of keys that have been pressed but not yet released, to allow providing the equivalent ReceivedCharacter data as was received for the press event.

Methods

impl Window

fn new(is_foreground: bool,
window_size: TypedSize2D<u32, ScreenPx>,
parent: Option<WindowID>)
-> Rc<Window>

fn platform_window(&self) -> WindowID

fn nested_window_resize(width: u32, height: u32)

fn gl_version() -> GlRequest

fn load_gl_functions(window_kind: &WindowKind)

fn handle_window_event(&self, event: Event) -> bool

fn toggle_modifier(&self, modifier: KeyModifiers)

fn scroll_window(&self, scroll_location: ScrollLocation, phase: TouchEventType)

Helper function to send a scroll event.

fn handle_mouse(&self,
button: MouseButton,
action: ElementState,
x: i32,
y: i32)

Helper function to handle a click

fn handle_next_event(&self) -> bool

fn wait_events(&self) -> Vec<WindowEvent>

unsafe fn set_nested_event_loop_listener(&self,
listener: *mut NestedEventLoopListener + 'static)

unsafe fn remove_nested_event_loop_listener(&self)

fn glutin_key_to_script_key(key: VirtualKeyCode) -> Result<Key, ()>

fn glutin_mods_to_script_mods(modifiers: KeyModifiers) -> KeyModifiers

fn platform_handle_key(&self, key: Key, mods: KeyModifiers)